2.0.0 (#253)
* bump version

* rename window to pane

* Add internal editor

* show chars left

* don't assume xdg-open, use open on iOS and start on Windows

* remove leaders and fix "switch"

* support multiple accounts

* update title on named feeds

* terminal-title=3, no top bar and no terminal title

* remove ini package

* add commands-in-new-pane and dynamic-timeline-name

* add delete-temp-files

* add lang to toots

* update docs

* update modules

* add clear-temp

* upgrade modules

* add gruvbox light

+++ ui/media.go
@@ -155,18 +155,22 @@ func openMediaType(tv *TutView, filenames []string, mediaType string) {
 	go func() {
 		for _, ext := range external {
 			exec.Command(ext.Name, ext.Args...).Run()
-			deleteFiles(ext.Filenames)
+			deleteFiles(tv, ext.Filenames)
 		}
 	}()
 	for _, term := range terminal {
 		openInTerminal(tv, term.Name, term.Args...)
-		deleteFiles(term.Filenames)
+		deleteFiles(tv, term.Filenames)
 	}
 }
 
-func deleteFiles(filenames []string) {
-	for _, filename := range filenames {
-		os.Remove(filename)
+func deleteFiles(tv *TutView, filenames []string) {
+	if tv.tut.Config.Media.DeleteTmpFiles {
+		for _, filename := range filenames {
+			os.Remove(filename)
+		}
+	} else {
+		tv.FileList = append(tv.FileList, filenames...)
 	}
 }
